Added support graceful data migrations.
Removed the exception thrown when decoding an object with a field missing. Decoding will now skip over missing fields silently. Classes should have default values declared in the default constructor to handle this case, which is a best practice anyway.

Added tests for going from an object with more fields to less fields and vica versa.

Signed-off-by: Patrick Hogan <pbhogan@gmail.com>